Extract test functions
