+++ example/App.js
@@ -31,6 +31,11 @@ import {
 	packOptionList
 } from './utils/businessLogic'
 import { generateMockData, generateMoreHistoricalData } from './utils/generateData'
+import {
+	testUpdateLastCandlestick,
+	testAddCandlesticksAtTheEnd,
+	testAddCandlesticksAtTheStart
+} from './utils/testUtils'
 
 
 const App = () => {
@@ -45,7 +50,6 @@ const App = () => {
 	const [klineData, setKlineData] = useState(generateMockData())
 	const [drawShouldContinue, setDrawShouldContinue] = useState(true)
 	const [optionList, setOptionList] = useState(null)
-	const [isLoadingNewData, setIsLoadingNewData] = useState(false)
 	const [lastDataLength, setLastDataLength] = useState(0)
 	const [currentScrollPosition, setCurrentScrollPosition] = useState(0)
 	const [showVolumeChart, setShowVolumeChart] = useState(true)
@@ -155,14 +159,13 @@ const App = () => {
 			klineData,
 			drawShouldContinue,
 			optionList,
-			isLoadingNewData,
 			lastDataLength,
 			currentScrollPosition,
 			showVolumeChart,
 			candleCornerRadius
 		}, shouldScrollToEnd)
 		setOptionListValue(newOptionList)
-	}, [klineData, selectedMainIndicator, selectedSubIndicator, showVolumeChart, isDarkTheme, selectedTimeType, selectedDrawTool, showIndicatorSelector, showTimeSelector, showDrawToolSelector, drawShouldContinue, optionList, isLoadingNewData, lastDataLength, currentScrollPosition, candleCornerRadius])
+	}, [klineData, selectedMainIndicator, selectedSubIndicator, showVolumeChart, isDarkTheme, selectedTimeType, selectedDrawTool, showIndicatorSelector, showTimeSelector, showDrawToolSelector, drawShouldContinue, optionList, lastDataLength, currentScrollPosition, candleCornerRadius])
 
 	// Reload K-line data and adjust scroll position to maintain current view
 	const reloadKLineDataWithScrollAdjustment = useCallback((addedDataCount) => {
@@ -188,7 +191,6 @@ const App = () => {
 			klineData,
 			drawShouldContinue,
 			optionList,
-			isLoadingNewData,
 			lastDataLength,
 			currentScrollPosition,
 			showVolumeChart,
@@ -209,7 +211,7 @@ const App = () => {
 		console.log(`Adjust scroll position: ${addedDataCount} data points, scroll distance: ${scrollAdjustment}px`)
 
 		setOptionListValue(newOptionList)
-	}, [klineData, selectedMainIndicator, selectedSubIndicator, showVolumeChart, isDarkTheme, selectedTimeType, selectedDrawTool, showIndicatorSelector, showTimeSelector, showDrawToolSelector, drawShouldContinue, optionList, isLoadingNewData, lastDataLength, currentScrollPosition, candleCornerRadius])
+	}, [klineData, selectedMainIndicator, selectedSubIndicator, showVolumeChart, isDarkTheme, selectedTimeType, selectedDrawTool, showIndicatorSelector, showTimeSelector, showDrawToolSelector, drawShouldContinue, optionList, lastDataLength, currentScrollPosition, candleCornerRadius])
 
 	// Set optionList property
 	const setOptionListValue = useCallback((optionList) => {
@@ -264,417 +266,36 @@ const App = () => {
 		}
 	}, [selectedDrawTool])
 
+
+	const handleTestAddCandlesticksAtTheStart = useCallback(() => {
+		console.log("handleTestAddCandlesticksAtTheStart called")
+		testAddCandlesticksAtTheStart(klineData, showVolumeChart, firstCandleTimeRef.current, (candlesticks) => {
+			kLineViewRef.current?.addCandlesticksAtTheStart(candlesticks)
+			firstCandleTimeRef.current = candlesticks[0].time
+		})
+	}, [klineData, showVolumeChart,kLineViewRef.current,firstCandleTimeRef.current])
+
 	// Handle new data loading triggered by left swipe
 	const handleScrollLeft = useCallback((event) => {
-		console.log('handleScrollLeft triggered, isLoadingNewData:', isLoadingNewData)
-		if (!isLoadingNewData) {
 			console.log('Loading 200 new historical candlesticks at start')
-			testAddCandlesticksAtTheStart()
-		} else {
-			console.log('Already loading data, skipping...')
-		}
-	}, [isLoadingNewData, testAddCandlesticksAtTheStart])
-
-	// Load more historical data
-	const loadMoreHistoricalData = useCallback(() => {
-		console.log("loadMoreHistoricalData called")
-		const currentData = klineData
-		const newHistoricalData = generateMoreHistoricalData(currentData, 200)
-		const combinedData = [...newHistoricalData, ...currentData]
-
-		console.log(`Loaded ${newHistoricalData.length} new historical K-line data points`)
-
-		// Calculate scroll offset adjustment needed to maintain current view
-		const addedDataCount = newHistoricalData.length
-
-		setKlineData(combinedData)
-		setLastDataLength(currentData.length)
-		setIsLoadingNewData(false)
-
-		// Reload data and maintain current view position
-		setTimeout(() => reloadKLineDataWithScrollAdjustment(addedDataCount), 0)
-	}, [klineData, reloadKLineDataWithScrollAdjustment])
-
-	// Test function to update last candlestick
-	const testUpdateLastCandlestick = useCallback(() => {
-		if (!kLineViewRef.current || klineData.length === 0) {
-			console.warn('No chart ref or data available')
-			return
-		}
-
-		const lastCandle = klineData[klineData.length - 1]
-
-		// Create a modified version of the last candlestick with random price changes
-		const priceChange = (0.01 - Math.random() * 0.02) * lastCandle.close // Random change between -1 and 1
-		const volumeChange = Math.random() * 0.1 + 1 // Random multiplier between 0.75 and 1.25
-
-		const newClose = Math.max(0.01, lastCandle.close + priceChange)
-		// Use vol field (native code expects 'vol' not 'volume')
-		const currentVolume = lastCandle.vol || 100000
-		const newVolume = Math.round(currentVolume * volumeChange)
-		// console.log('Volume calculation:', { currentVolume, volumeChange, newVolume })
-
-		// Calculate MA indicators for the updated candlestick
-		const currentIndex = klineData.length - 1
-
-		// Helper function to safely get volume value
-		const getSafeVolume = (item) => {
-			const vol = item.vol || item.volume
-			return isNaN(vol) || !isFinite(vol) ? 100000 : vol
-		}
-
-		// Calculate Volume MA5
-		let volumeMa5 = newVolume
-		if (currentIndex >= 4) {
-			let sum = newVolume
-			// Get the previous 4 candles (not including the current one being updated)
-			for (let j = Math.max(0, currentIndex - 4); j < currentIndex; j++) {
-				sum += getSafeVolume(klineData[j])
-			}
-			volumeMa5 = sum / 5
-			// console.log('Volume MA5 calculation:', { newVolume, sum, volumeMa5, currentIndex })
-		}
-
-		// Calculate Volume MA10
-		let volumeMa10 = newVolume
-		if (currentIndex >= 9) {
-			let sum = newVolume
-			// Get the previous 9 candles (not including the current one being updated)
-			for (let j = Math.max(0, currentIndex - 9); j < currentIndex; j++) {
-				sum += getSafeVolume(klineData[j])
-			}
-			volumeMa10 = sum / 10
-			// console.log('Volume MA10 calculation:', { newVolume, sum, volumeMa10, currentIndex })
-		}
-
-		// Calculate price MA indicators
-		let ma5 = newClose
-		if (currentIndex >= 4) {
-			let sum = newClose
-			for (let j = Math.max(0, currentIndex - 4); j < currentIndex; j++) {
-				sum += klineData[j].close
-			}
-			ma5 = sum / 5
-		}
-
-		let ma10 = newClose
-		if (currentIndex >= 9) {
-			let sum = newClose
-			for (let j = Math.max(0, currentIndex - 9); j < currentIndex; j++) {
-				sum += klineData[j].close
-			}
-			ma10 = sum / 10
-		}
-
-		let ma20 = newClose
-		if (currentIndex >= 19) {
-			let sum = newClose
-			for (let j = Math.max(0, currentIndex - 19); j < currentIndex; j++) {
-				sum += klineData[j].close
-			}
-			ma20 = sum / 20
-		}
-
-		// Ensure all values are valid numbers
-		const safeValue = (val, fallback = 0) => isNaN(val) || !isFinite(val) ? fallback : val
-
-		// Remove the volume field to avoid confusion (we use vol)
-		const { volume, ...lastCandleWithoutVolume } = lastCandle
-
-		const updatedCandle = {
-			...lastCandleWithoutVolume,
-			close: newClose,
-			high: Math.max(lastCandle.high, newClose + Math.abs(priceChange) * 0.5),
-			low: Math.min(lastCandle.low, newClose - Math.abs(priceChange) * 0.5),
-			vol: newVolume,
-			// Ensure required fields are present
-			id: lastCandle.id || lastCandle.time,
-			dateString: lastCandle.dateString || new Date(lastCandle.time).toISOString(),
-			// Add calculated indicator lists
-			maList: [
-				{ title: '5', value: safeValue(ma5, newClose), selected: true, index: 0 },
-				{ title: '10', value: safeValue(ma10, newClose), selected: true, index: 1 },
-				{ title: '20', value: safeValue(ma20, newClose), selected: true, index: 2 }
-			],
-			maVolumeList: [
-				{ title: '5', value: safeValue(volumeMa5, 100000), selected: showVolumeChart, index: 0 },
-				{ title: '10', value: safeValue(volumeMa10, 100000), selected: showVolumeChart, index: 1 }
-			],
-			rsiList: lastCandle.rsiList || [],
-			wrList: lastCandle.wrList || [],
-			selectedItemList: lastCandle.selectedItemList || [],
-			// Add placeholder values for BOLL and KDJ indicators
-			bollMb: newClose,  // Middle band (moving average)
-			bollUp: newClose * 1.02,  // Upper band (simplified)
-			bollDn: newClose * 0.98,  // Lower band (simplified)
-			kdjK: 50,  // Placeholder K value
-			kdjD: 50,  // Placeholder D value
-			kdjJ: 50   // Placeholder J value
-		}
-
-		console.log('Updating last candlestick:', updatedCandle)
-		// console.log('Volume MA values being sent:', {
-		//	volumeMa5: updatedCandle.maVolumeList[0].value,
-		//	volumeMa10: updatedCandle.maVolumeList[1].value,
-		//	newVolume: updatedCandle.vol
-		// })
-
-		// Call the native method directly
-		kLineViewRef.current.updateLastCandlestick(updatedCandle)
-	}, [klineData, showVolumeChart])
-
-	// Test function to add new candlesticks at the end
-	const testAddCandlesticksAtTheEnd = useCallback(() => {
-		if (!kLineViewRef.current || klineData.length === 0) {
-			console.warn('No chart ref or data available')
-			return
-		}
+			handleTestAddCandlesticksAtTheStart()
+	}, [handleTestAddCandlesticksAtTheStart])
 
 
-		const lastCandle = klineData[klineData.length - 1]
-		const numberOfNewCandles = 200 // Add 200 new candlesticks
+	// Wrapper functions for test utilities
+	const handleTestUpdateLastCandlestick = useCallback(() => {
+		testUpdateLastCandlestick(klineData, showVolumeChart, (candlestick) => {
+			kLineViewRef.current?.updateLastCandlestick(candlestick)
+		})
+	}, [klineData, showVolumeChart,kLineViewRef.current])
 
-		// Generate new candlesticks with indicators
-		const newCandlesticks = []
-		for (let i = 1; i <= numberOfNewCandles; i++) {
-			const timeIncrement = 60000 * i // 1 minute intervals
-			const basePrice = lastCandle.close
-			const priceVariation = (Math.random() - 0.5) * basePrice * 0.02 // ±2% variation
+	const handleTestAddCandlesticksAtTheEnd = useCallback(() => {
+		testAddCandlesticksAtTheEnd(klineData, showVolumeChart, (candlesticks) => {
+			kLineViewRef.current?.addCandlesticksAtTheEnd(candlesticks)
+		})
+	}, [klineData, showVolumeChart,kLineViewRef.current])
 
-			const open = Math.max(0.01, basePrice + (Math.random() - 0.5) * basePrice * 0.01)
-			const close = Math.max(0.01, basePrice + priceVariation)
-			const high = Math.max(open, close) + Math.random() * basePrice * 0.005
-			const low = Math.min(open, close) - Math.random() * basePrice * 0.005
-			const volume = Math.round(lastCandle.vol * (0.5 + Math.random()))
 
-			// Calculate MA indicators based on historical data
-			const tempAllData = [...klineData]
-			const currentIndex = tempAllData.length + i - 1
-
-			// Calculate MA5
-			let ma5 = close
-			if (currentIndex >= 4) {
-				let sum = close
-				for (let j = Math.max(0, tempAllData.length - 4 + i - 1); j < tempAllData.length; j++) {
-					sum += tempAllData[j].close
-				}
-				ma5 = sum / 5
-			}
-
-			// Calculate MA10
-			let ma10 = close
-			if (currentIndex >= 9) {
-				let sum = close
-				for (let j = Math.max(0, tempAllData.length - 9 + i - 1); j < tempAllData.length; j++) {
-					sum += tempAllData[j].close
-				}
-				ma10 = sum / 10
-			}
-
-			// Calculate MA20
-			let ma20 = close
-			if (currentIndex >= 19) {
-				let sum = close
-				for (let j = Math.max(0, tempAllData.length - 19 + i - 1); j < tempAllData.length; j++) {
-					sum += tempAllData[j].close
-				}
-				ma20 = sum / 20
-			}
-
-			// Helper function to safely get volume value
-			const getSafeVolume = (item) => {
-				const vol = item.vol || item.volume
-				return isNaN(vol) || !isFinite(vol) ? 100000 : vol
-			}
-
-			// Calculate Volume MA5
-			let volumeMa5 = volume
-			if (currentIndex >= 4) {
-				let sum = volume
-				for (let j = Math.max(0, tempAllData.length - 4 + i - 1); j < tempAllData.length; j++) {
-					sum += getSafeVolume(tempAllData[j])
-				}
-				volumeMa5 = sum / 5
-			}
-
-			// Calculate Volume MA10
-			let volumeMa10 = volume
-			if (currentIndex >= 9) {
-				let sum = volume
-				for (let j = Math.max(0, tempAllData.length - 9 + i - 1); j < tempAllData.length; j++) {
-					sum += getSafeVolume(tempAllData[j])
-				}
-				volumeMa10 = sum / 10
-			}
-
-			// Ensure all values are valid numbers
-			const safeValue = (val, fallback = 0) => isNaN(val) || !isFinite(val) ? fallback : val
-
-			const newCandle = {
-				time: lastCandle.time + timeIncrement,
-				open: parseFloat(open.toFixed(2)),
-				high: parseFloat(high.toFixed(2)),
-				low: parseFloat(low.toFixed(2)),
-				close: parseFloat(close.toFixed(2)),
-				vol: safeValue(volume, 100000), // Fallback to reasonable volume
-				id: lastCandle.time + timeIncrement,
-				dateString: formatTime(lastCandle.time + timeIncrement, 'MM-DD HH:mm'),
-				// Add indicator lists
-				maList: [
-					{ title: '5', value: safeValue(ma5, close), selected: true, index: 0 },
-					{ title: '10', value: safeValue(ma10, close), selected: true, index: 1 },
-					{ title: '20', value: safeValue(ma20, close), selected: true, index: 2 }
-				],
-				maVolumeList: [
-					{ title: '5', value: safeValue(volumeMa5, 100000), selected: showVolumeChart, index: 0 },
-					{ title: '10', value: safeValue(volumeMa10, 100000), selected: showVolumeChart, index: 1 }
-				],
-				rsiList: lastCandle.rsiList || [],
-				wrList: lastCandle.wrList || [],
-				selectedItemList: lastCandle.selectedItemList || [],
-				// Add placeholder values for BOLL and KDJ indicators
-				bollMb: close,  // Middle band (moving average)
-				bollUp: close * 1.02,  // Upper band (simplified)
-				bollDn: close * 0.98,  // Lower band (simplified)
-				kdjK: 50,  // Placeholder K value
-				kdjD: 50,  // Placeholder D value
-				kdjJ: 50   // Placeholder J value
-			}
-
-			newCandlesticks.push(newCandle)
-			// Add to temp array for next iteration calculations
-			tempAllData.push(newCandle)
-		}
-
-		console.log('Adding', numberOfNewCandles, 'new candlesticks at the end:', newCandlesticks)
-
-		// Update local state for future reference
-		// setKlineData(prev => [...prev, ...newCandlesticks])
-
-		// Call the native method directly
-		kLineViewRef.current.addCandlesticksAtTheEnd(newCandlesticks)
-	}, [klineData])
-
-	// Test function to add new candlesticks at the start
-	const testAddCandlesticksAtTheStart = useCallback(() => {
-		if (!kLineViewRef.current || klineData.length === 0) {
-			console.warn('No chart ref or data available')
-			return
-		}
-
-		if (isLoadingNewData) {
-			console.log('Already loading data, skipping...')
-			return
-		}
-
-		// setIsLoadingNewData(true)
-		console.log('Starting to load 200 new historical candlesticks')
-
-		const numberOfNewCandles = 200 // Load 200 candlesticks
-		const newCandlesticks = []
-		const firstCandle = klineData[0]
-		const timeIncrement = -1 * 60 * 1000 // Go backward 1 minute
-
-		// Create temp array for calculations (prepend to existing data)
-		const tempAllData = [...klineData]
-
-		for (let i = 0; i < numberOfNewCandles; i++) {
-			// Calculate timestamp going backwards in time (oldest candles first in our generation)
-			// i=0 gives us the oldest candle (200 minutes ago), i=199 gives us the newest (1 minute ago)
-			const minutesBack = numberOfNewCandles - i // 200, 199, 198, ..., 1
-			const timestamp = firstCandleTimeRef.current - (minutesBack * 60 * 1000)
-
-			// Generate price data going backward in time
-			const basePrice = firstCandle.open
-			const priceVariation1 = (Math.random() - 0.5) * basePrice * 0.02
-			const priceVariation2 = (Math.random() - 0.5) * basePrice * 0.02
-			const open = Math.max(0.01, basePrice + priceVariation1)
-			const close = Math.max(0.01, basePrice + priceVariation2)
-			const high = Math.max(open, close) + Math.random() * basePrice * 0.005
-			const low = Math.min(open, close) - Math.random() * basePrice * 0.005
-			const volume = Math.round(firstCandle.vol * (0.5 + Math.random()))
-
-			// Calculate MA indicators based on position in the prepended data
-			const currentIndex = numberOfNewCandles - 1 - i // Position in the new data
-
-			// Helper function to safely get volume value
-			const getSafeVolume = (item) => {
-				const vol = item.vol || item.volume
-				return isNaN(vol) || !isFinite(vol) ? 100000 : vol
-			}
-
-			// Calculate Volume MA5 (for the start, use current volume as base)
-			let volumeMa5 = volume
-			// Since we're at the start, use simple fallback
-
-			// Calculate Volume MA10
-			let volumeMa10 = volume
-
-			// Calculate MA5
-			let ma5 = close
-
-			// Calculate MA10
-			let ma10 = close
-
-			// Calculate MA20
-			let ma20 = close
-
-			// Ensure all values are valid numbers
-			const safeValue = (val, fallback = 0) => isNaN(val) || !isFinite(val) ? fallback : val
-
-			const newCandle = {
-				time: timestamp,
-				open: parseFloat(open.toFixed(2)),
-				high: parseFloat(high.toFixed(2)),
-				low: parseFloat(low.toFixed(2)),
-				close: parseFloat(close.toFixed(2)) - 200,
-				vol: safeValue(volume, 100000), // Fallback to reasonable volume
-				id: timestamp,
-				dateString: formatTime(timestamp, 'MM-DD HH:mm'),
-				// Add indicator lists
-				maList: [
-					{ title: '5', value: safeValue(ma5, close), selected: true, index: 0 },
-					{ title: '10', value: safeValue(ma10, close), selected: true, index: 1 },
-					{ title: '20', value: safeValue(ma20, close), selected: true, index: 2 }
-				],
-				maVolumeList: [
-					{ title: '5', value: safeValue(volumeMa5, 100000), selected: showVolumeChart, index: 0 },
-					{ title: '10', value: safeValue(volumeMa10, 100000), selected: showVolumeChart, index: 1 }
-				],
-				rsiList: firstCandle.rsiList || [],
-				wrList: firstCandle.wrList || [],
-				selectedItemList: firstCandle.selectedItemList || [],
-				// Add placeholder values for BOLL and KDJ indicators
-				bollMb: close,  // Middle band (moving average)
-				bollUp: close * 1.02,  // Upper band (simplified)
-				bollDn: close * 0.98,  // Lower band (simplified)
-				kdjK: 50,  // Placeholder K value
-				kdjD: 50,  // Placeholder D value
-				kdjJ: 50   // Placeholder J value
-			}
-
-			newCandlesticks.push(newCandle) // Add in chronological order (oldest to newest)
-		}
-		firstCandleTimeRef.current = newCandlesticks[0].time
-
-		console.log('Adding', numberOfNewCandles, 'new candlesticks at the start:')
-		console.log("First historical candle (oldest):", newCandlesticks[0])
-		console.log("Last historical candle (newest):", newCandlesticks[newCandlesticks.length - 1])
-		console.log("Previous first candle:", firstCandle)
-		console.log("Timestamp comparison:")
-		console.log("  Historical oldest:", new Date(newCandlesticks[0].time).toLocaleString())
-		console.log("  Historical newest:", new Date(newCandlesticks[newCandlesticks.length - 1].time).toLocaleString())
-		console.log("  Previous first:", new Date(firstCandle.time).toLocaleString())
-
-		// Call the native method directly
-		kLineViewRef.current.addCandlesticksAtTheStart(newCandlesticks)
-
-		// // Reset loading state after a delay to allow chart to update
-		// setTimeout(() => {
-		// 	setIsLoadingNewData(false)
-		// 	console.log('Loading state reset, ready for next load')
-		// }, 1000)
-	}, [klineData, showVolumeChart, isLoadingNewData])
 
 	const renderKLineChart = useCallback((styles) => {
 		const directRender = (
@@ -737,9 +358,9 @@ const App = () => {
 				theme={theme}
 				isDarkTheme={isDarkTheme}
 				onToggleTheme={toggleTheme}
-				onTestUpdate={testUpdateLastCandlestick}
-				onTestAddCandles={testAddCandlesticksAtTheEnd}
-				onTestAddCandlesAtStart={testAddCandlesticksAtTheStart}
+				onTestUpdate={handleTestUpdateLastCandlestick}
+				onTestAddCandles={handleTestAddCandlesticksAtTheEnd}
+				onTestAddCandlesAtStart={handleTestAddCandlesticksAtTheStart}
 			/>
 
 			{/* K-line chart */}
